requirements.txt
torch==2.2.0
torchvision==0.17.0
opencv-python==4.8.0.74
matplotlib 
jupyter
notebook 
hydra-core
scikit-image
numpy<2.0
tqdm
tensorboard
omegaconf
numpy 
Pillow
piqa
datetime
timm
einops
transformers
datasets
grad-cam